fix: UserProvider를 이동하여 전역에서 사용자 상태 초기화
